Updated to C++17 and applied consistent formatting
This code was still using some old C++11 features like std::bind instead
of lambdas and a custom make_unique (since make_unique wasn't added
until later). It also had a random mixture of indentation and code
styles as a result of two different people working on it. I updated
everything to a more modern standard, and ran clang-format on most of
the files (except for the ones that are copied-and-pasted libraries,
like SuffixTree.cpp).